Rename `es` jobs to `elasticsearch`
I noticed that Elasticsearch jobs don't follow the naming convention of the
other imported files in this repo - namely, that if they refer to a
particular product or software, they use the full name rather than some
truncated version.

Functionally, there is no difference. Aesthetically, and logically, I prefer
this change. It neaten things out, and it ensures we're following previous
standards we've set ourselves.
